+++ back-end/src/main/java/com/tfg/back/service/serviceImpl/DoctorServiceImpl.java
@@ -11,6 +11,7 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
+import java.util.Optional;
 
 @Service
 public class DoctorServiceImpl implements DoctorService {
@@ -33,7 +34,8 @@ public class DoctorServiceImpl implements DoctorService {
 
     @Override
     public Doctor getDoctor(Long id) {
-        return doctorRepository.findById(id).get();
+        Optional<Doctor> doctor = doctorRepository.findById(id);
+        return doctor.orElse(null);
     }
 
     @Override
@@ -42,7 +44,13 @@ public class DoctorServiceImpl implements DoctorService {
     }
 
     @Override
-    public void deleteDoctor(Long id) {
-        doctorRepository.deleteById(id);
+    public boolean deleteDoctor(Long id) {
+        boolean deleted = false;
+        boolean exists = doctorRepository.existsById(id);
+        if (exists){
+            doctorRepository.deleteById(id);
+            deleted = true;
+        }
+        return deleted;
     }
 }
